+++ src/cmd/versions.rs
@@ -3,7 +3,7 @@ use crate::output::Output;
 use crate::Result;
 use std::process::ExitCode;
 
-pub fn versions(app_name: &str, amount: u8, output: &dyn Output) -> Result<ExitCode> {
+pub fn versions(app_name: &str, amount: usize, output: &dyn Output) -> Result<ExitCode> {
     let apps = &apps::all();
     let app = apps.lookup(app_name)?;
     let versions = app.versions(amount, output)?;
